// src/models/car.ts
class Car {
    baalut: string = "";
    degem_cd: number = 0;
    degem_manoa: string = "";
    degem_nm: string = "";
    horaat_rishum: number = 0;
    kinuy_mishari: string = "";
    kvutzat_zihum: number = 0;
    misgeret: string = "";
    mispar_rechev: number = 0;
    mivchan_acharon_dt: string = "";
    moed_aliya_lakvish: string = "";
    ramat_eivzur_betihuty: string = "";
    ramat_gimur: string = "";
    rank: number = 0;
    shnat_yitzur: number = 0;
    sug_degem: string = "";
    sug_delek_nm: string = "";
    tokef_dt:string = "";
    tozeret_cd:number = 0;
    tozeret_nm:string = "";
    tzeva_cd:number = 0;
    tzeva_rechev:string = "";
    zmig_ahori:string = "";
    zmig_kidmi:string = "";
    _id:number = 0;
}

export default Car;